' I have test the gui, Thanks for your work most parts are great! I fix 3 parts to make the GUI complete. First，In the function enroled subect in student system, If thereis no subject in the enrollment list, we can only see the prompt message without a return button in the interface. I add a button when theres no enrollment subject in the list.
Second, Get Students by Grade in Admin Sys. I set a scroll wheel function for this part.
If there are several students with several subjects need to be shown in the list,
we cannot list all info in the screen and we cannot find and use the back button in the bottom.
I change the code with the scroll wheel to show all info and could find the back button.

Third part is the same as second in the Categorize by Pass\fail in the admin sys.
I did the same as in the second part.
'